rename `SinglyLinkList` to `SinglyLinkedList`
